chain: Don't allow balance to go below zero
This shouldn't happen in most cases anwyay, since we're already
detecting if the user has enough money to spend and are not
spending negative money. However, we also want to make sure that
the genesis block can give money to itself and so this requires
that the transaction is not necessarily zero-sum.