chain: Better concecptually separate Block and Transaction
A block is now essentially just an arbitrary payload with a
nonce and a hash, computed with reference to the parent hash.

If we need to mutate the contents of a block, you first need to convert
it to whatever it is that you want to modify, and then reserialize the
payload. The Transaction class provides a convenience method
withTransaction just for that purpose..