chian: Move the mining process to a separate thread
A bunch of change fallout that happened because of this. All the
validation needs to happen just before a block is mined, which means
that we can no longer validate synchronously. Everything needs to
happen async and we needed to add a bunch of mechanisms to allow
the producer and consumer to synchronise states with each other.