chain: Separate out AsynchronouslyMutableLedger and BlockMiner
The mining process is somewhat orthogonal to the chain and ledger
structure itself. The wallet will want to utilize both of these
in order to validate the structure of the data structure itself,
but the miner will need the Async* variants in order to append new
transactions. That happens asynchronously and there's a lot more
to set up and take care of.